requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.8
django-environ==0.11.2
django-ratelimit==4.1.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.4.2
idna==3.6
pymongo==4.6.1
python-decouple==3.8
pytz==2023.3.post1
ratelimit==2.2.1
requests==2.31.0
sqlparse==0.4.4
typing-extensions==4.8.0
urllib3==2.1.0
